Lesson 40 - 41 - Displaying Vocations
List each vocation in a scrollable box. Experimented with doing a horizontal scroll and only showing the image, since the image has all the text info we need anyway. Likely will switch to this in the future instead of the ugly verticle scroll. The only reason I'm not doing it now is because I want to see how we style the selected vocation... some kind of border color I assume.