Lesson 45 - Navigation
Shaun used Navigator.push and defined a MaterialPageRoute - I didn't really like that approach, I prefer central route management. So I updated main.dart with a route table and use Navigator.pushNamed, then pass in the name of the route.

Downside of this approach would be custom transitions, which might be covered in lesson 75-83